# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Book Store API",
...
    "contact": {
      "name": "Book Store Support"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:60385/"
    }
  ],
  "paths": {
...
    "/api/publishers": {
      "get": {
        "tags": [
          "Publishers"
        ],
        "summary": "Get all publishers",
        "operationId": "GetPublishers",
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "api-version",
            "in": "header",
            "description": "API version. Example: `1.0`"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred language for localized content (configurable via LocalizationOptions)"
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Correlation ID for tracking requests across the system"
          },
          {
            "name": "X-Causation-ID",
            "in": "header",
            "description": "Causation ID linking this request to its trigger"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublisherDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/publishers/{id}": {
      "get": {
        "tags": [
          "Publishers"
        ],
        "summary": "Get publisher by ID",
        "operationId": "GetPublisher",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "api-version",
            "in": "header",
            "description": "API version. Example: `1.0`"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred language for localized content (configurable via LocalizationOptions)"
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Correlation ID for tracking requests across the system"
          },
          {
            "name": "X-Causation-ID",
            "in": "header",
            "description": "Causation ID linking this request to its trigger"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/admin/publishers": {
      "post": {
        "tags": [
          "Admin - Publishers"
        ],
        "summary": "Create a new publisher",
        "operationId": "CreatePublisher",
        "parameters": [
          {
            "name": "api-version",
            "in": "header",
            "description": "API version. Example: `1.0`"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred language for localized content (configurable via LocalizationOptions)"
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Correlation ID for tracking requests across the system"
          },
          {
            "name": "X-Causation-ID",
            "in": "header",
            "description": "Causation ID linking this request to its trigger"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePublisherRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/publishers/{id}": {
      "put": {
        "tags": [
          "Admin - Publishers"
        ],
        "summary": "Update a publisher",
        "operationId": "UpdatePublisher",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "api-version",
            "in": "header",
            "description": "API version. Example: `1.0`"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred language for localized content (configurable via LocalizationOptions)"
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Correlation ID for tracking requests across the system"
          },
          {
            "name": "X-Causation-ID",
            "in": "header",
            "description": "Causation ID linking this request to its trigger"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePublisherRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Publishers"
        ],
        "summary": "Delete a publisher",
        "operationId": "SoftDeletePublisher",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "api-version",
            "in": "header",
            "description": "API version. Example: `1.0`"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred language for localized content (configurable via LocalizationOptions)"
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Correlation ID for tracking requests across the system"
          },
          {
            "name": "X-Causation-ID",
            "in": "header",
            "description": "Causation ID linking this request to its trigger"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePublisherRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "PublisherDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UpdatePublisherRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      }
    }
  },
...
}